Move sample User Agent generation into the same module as UA generation
